RouletteWheel tests are passing

+++ src/tests/java/picard/analysis/TheoreticalSensitivityTest.java
@@ -25,7 +25,7 @@ public class TheoreticalSensitivityTest {
 
         //test the sums of this deterministic wheel: a sum of n 1's equals n
         List<ArrayList<Integer>> deterministicSums = deterministicWheel.sampleCumulativeSums(10, 1);
-        for (int n = 0; n < 10; n++) Assert.assertEquals(deterministicSums.get(n).get(0), n);
+        for (int n = 0; n < 10; n++) Assert.assertEquals(deterministicSums.get(n).get(0), (Integer) n);
 
         //test that an unfair coin w/ probability 1/4 of heads gives roughly 1/4 heads
         double p = 0.25;
